Add IncomingTransactionHelper
Add EtherscanRemoteTransactionSource.
Add Etherscan functions.
Move common types to separate file.
