+++ packages/transaction-controller/src/index.ts
@@ -1,2 +1,3 @@
 export * from './TransactionController';
 export { isEIP1559Transaction } from './utils';
+export * from './types';
