+++ packages/transaction-controller/src/types.ts
@@ -1,0 +1,117 @@
+import { NetworkType } from '@metamask/controller-utils';
+import { Hex } from '@metamask/utils';
+
+/**
+ * @type TransactionMeta
+ *
+ * TransactionMeta representation
+ * @property error - Synthesized error information for failed transactions
+ * @property id - Generated UUID associated with this transaction
+ * @property networkID - Network code as per EIP-155 for this transaction
+ * @property origin - Origin this transaction was sent from
+ * @property deviceConfirmedOn - string to indicate what device the transaction was confirmed
+ * @property rawTransaction - Hex representation of the underlying transaction
+ * @property status - String status of this transaction
+ * @property time - Timestamp associated with this transaction
+ * @property toSmartContract - Whether transaction recipient is a smart contract
+ * @property transaction - Underlying Transaction object
+ * @property transactionHash - Hash of a successful transaction
+ * @property blockNumber - Number of the block where the transaction has been included
+ */
+export type TransactionMeta =
+  | ({
+      status: Exclude<TransactionStatus, TransactionStatus.failed>;
+    } & TransactionMetaBase)
+  | ({ status: TransactionStatus.failed; error: Error } & TransactionMetaBase);
+
+type TransactionMetaBase = {
+  isTransfer?: boolean;
+  transferInformation?: {
+    symbol: string;
+    contractAddress: string;
+    decimals: number;
+  };
+  id: string;
+  networkID?: string;
+  chainId?: Hex;
+  origin?: string;
+  rawTransaction?: string;
+  time: number;
+  toSmartContract?: boolean;
+  transaction: Transaction;
+  transactionHash?: string;
+  blockNumber?: string;
+  deviceConfirmedOn?: WalletDevice;
+  verifiedOnBlockchain?: boolean;
+};
+
+/**
+ * The status of the transaction. Each status represents the state of the transaction internally
+ * in the wallet. Some of these correspond with the state of the transaction on the network, but
+ * some are wallet-specific.
+ */
+export enum TransactionStatus {
+  approved = 'approved',
+  cancelled = 'cancelled',
+  confirmed = 'confirmed',
+  failed = 'failed',
+  rejected = 'rejected',
+  signed = 'signed',
+  submitted = 'submitted',
+  unapproved = 'unapproved',
+}
+
+/**
+ * Options for wallet device.
+ */
+export enum WalletDevice {
+  MM_MOBILE = 'metamask_mobile',
+  MM_EXTENSION = 'metamask_extension',
+  OTHER = 'other_device',
+}
+
+/**
+ * @type Transaction
+ *
+ * Transaction representation
+ * @property chainId - Network ID as per EIP-155
+ * @property data - Data to pass with this transaction
+ * @property from - Address to send this transaction from
+ * @property gas - Gas to send with this transaction
+ * @property gasPrice - Price of gas with this transaction
+ * @property gasUsed - Gas used in the transaction
+ * @property nonce - Unique number to prevent replay attacks
+ * @property to - Address to send this transaction to
+ * @property value - Value associated with this transaction
+ */
+export interface Transaction {
+  chainId?: Hex;
+  data?: string;
+  from: string;
+  gas?: string;
+  gasPrice?: string;
+  gasUsed?: string;
+  nonce?: string;
+  to?: string;
+  value?: string;
+  maxFeePerGas?: string;
+  maxPriorityFeePerGas?: string;
+  estimatedBaseFee?: string;
+  estimateGasError?: string;
+}
+
+export interface RemoteTransactionSourceRequest {
+  address: string;
+  networkType: NetworkType;
+  limit: number;
+  currentChainId: Hex;
+  currentNetworkId: string;
+  fromBlock?: string;
+  apiKey?: string;
+}
+
+export interface RemoteTransactionSource {
+  fetchTransactions: (
+    request: RemoteTransactionSourceRequest,
+  ) => Promise<TransactionMeta[]>;
+}
