+++ packages/transaction-controller/src/utils.test.ts
@@ -1,14 +1,12 @@
 import type { Transaction as NonceTrackerTransaction } from 'nonce-tracker/dist/NonceTracker';
 
-import { TransactionStatus } from './TransactionController';
 import type {
-  TransactionMeta,
-  Transaction,
   GasPriceValue,
   FeeMarketEIP1559Values,
 } from './TransactionController';
 import * as util from './utils';
 import { getAndFormatTransactionsForNonceTracker } from './utils';
+import { Transaction, TransactionMeta, TransactionStatus } from './types';
 
 const MAX_FEE_PER_GAS = 'maxFeePerGas';
 const MAX_PRIORITY_FEE_PER_GAS = 'maxPriorityFeePerGas';
@@ -17,44 +15,6 @@ const FAIL = 'lol';
 const PASS = '0x1';
 
 describe('utils', () => {
-  describe('getEtherscanApiUrl', () => {
-    const networkType = 'mainnet';
-    const address = '0xC7D3BFDeA106B446Cf9f2Db354D496e6Dd8b2525';
-    const action = 'txlist';
-
-    it('should return a correctly structured url', () => {
-      const url = util.getEtherscanApiUrl(networkType, { address, action });
-      expect(url.indexOf(`&action=${action}`)).toBeGreaterThan(0);
-    });
-
-    it('should return a correctly structured url with from block', () => {
-      const fromBlock = 'xxxxxx';
-      const url = util.getEtherscanApiUrl(networkType, {
-        address,
-        action,
-        startBlock: fromBlock,
-      });
-      expect(url.indexOf(`&startBlock=${fromBlock}`)).toBeGreaterThan(0);
-    });
-
-    it('should return a correctly structured url with testnet subdomain', () => {
-      const goerli = 'goerli';
-      const url = util.getEtherscanApiUrl(goerli, { address, action });
-      expect(url.indexOf(`https://api-${goerli}`)).toBe(0);
-    });
-
-    it('should return a correctly structured url with apiKey', () => {
-      const apiKey = 'xxxxxx';
-      const url = util.getEtherscanApiUrl(networkType, {
-        address,
-        action,
-        startBlock: 'xxxxxx',
-        apikey: apiKey,
-      });
-      expect(url.indexOf(`&apikey=${apiKey}`)).toBeGreaterThan(0);
-    });
-  });
-
   it('normalizeTransaction', () => {
     const normalized = util.normalizeTransaction({
       data: 'data',
